--- src/components/BoardBar/index.js
@@ -1,1 +1,0 @@
-export { default as BoardBar } from './BoardBar';
