--- src/components/DropDownMenu/index.js
@@ -1,1 +1,0 @@
-export { default as DropDownMenu } from './DropDownMenu';
