--- src/components/NoticeBar/index.js
@@ -1,1 +1,0 @@
-export { default as NoticeBar } from './NoticeBar';
