--- src/components/PostBar/index.js
@@ -1,1 +1,0 @@
-export { default as PostBar } from './PostBar';
