+++ src/components/index.js
@@ -1,18 +1,14 @@
-export * from './BoardBar';
 export * from './BoardCard';
 export * from './Card';
 export * from './Carousel';
 export * from './Comment';
 export * from './Filter';
-export * from './DropDownMenu';
 export * from './Fieldset';
 export * from './Filter';
 export * from './Input';
 export * from './InputBar';
 export * from './ListHeader';
 export * from './MainPageListItem';
-export * from './NoticeBar';
-export * from './PostBar';
 export * from './PopUp';
 export * from './ReviewContentItem';
 export * from './ReviewDownload';
