+++ src/constants/index.js
@@ -3,6 +3,5 @@ export * from './admin.js';
 export * from './align.js';
 export * from './dropDown.js';
 export * from './examReview.js';
-export * from './postCategories.js';
 export * from './search.js';
 export * from './seconds.js';
