+++ src/feature/board/component/NoticeBar/NoticeBar.jsx
