+++ src/feature/board/component/PostBar/PostBar.jsx
