+++ src/feature/board/component/PostBar/PostBar.module.css
