+++ src/feature/board/component/index.js
@@ -1,0 +1,4 @@
+export { default as BoardBar } from './BoardBar/BoardBar';
+export { default as DropDownMenu } from './DropDownMenu/DropDownMenu';
+export { default as NoticeBar } from './NoticeBar/NoticeBar';
+export { default as PostBar } from './PostBar/PostBar';
