+++ src/pages/MyPage/pages/ActivityPage/components/Posts/Posts.jsx
@@ -9,7 +9,7 @@ import {
 } from '@/shared/lib';
 import { STALE_TIME } from '@/shared/constant';
 
-import { PostBar } from '@/components';
+import { PostBar } from '@/feature/board/component';
 import frustratedWomanIllustration from '@/assets/images/frustratedWoman.svg';
 
 import styles from './Posts.module.css';
