+++ src/pages/PostPage/PostWritePage/PostWritePage.jsx
@@ -15,7 +15,7 @@ import { formattedNowTime, getBoard } from '@/shared/lib';
 import { BOARD_MENUS, QUERY_KEY, ROLE, TOAST } from '@/shared/constant';
 
 import { postPost } from '@/apis';
-import { DropDownMenu } from '@/components';
+import { DropDownMenu } from '@/feature/board/component';
 
 import styles from './PostWritePage.module.css';
