+++ src/pages/SearchPage/components/SearchResults/SearchResults.jsx
@@ -10,7 +10,7 @@ import {
 } from '@/shared/lib';
 import { BOARDS } from '@/shared/constant';
 
-import { PostBar } from '@/components';
+import { PostBar } from '@/feature/board/component';
 
 import styles from './SearchResults.module.css';
